day11

+++ Day11.cs
@@ -12,7 +12,7 @@ public class Day11
   [InlineData("Day11.Sample", 25, 55312)]
   [InlineData("Day11", 25, 202019L)]
   [InlineData("Day11", 75, 239321955280205L)]
-  public void Part1(string file, int blinks, int expected)
+  public void Part1(string file, int blinks, long expected)
   {
     var input = FormatInput(AoCLoader.LoadLines(file));
     input.Select(it => NumberOfStonesAfterBlinking(it, blinks)).Sum().Should().Be(expected);
